Extra cell components from TokensTableRow
