+++ frontend/src/lib/components/tokens/TokensTable/TokenTitleCell.svelte
@@ -1,0 +1,42 @@
+<script lang="ts">
+  import type { UserTokenData, UserTokenLoading } from "$lib/types/tokens-page";
+  import Logo from "../../ui/Logo.svelte";
+  import { nonNullish } from "@dfinity/utils";
+
+  export let userTokenData: UserTokenData | UserTokenLoading;
+</script>
+
+<div class="title-logo-wrapper">
+  <Logo
+    src={userTokenData.logo}
+    alt={userTokenData.title}
+    size="medium"
+    framed
+  />
+  <div class="title-wrapper">
+    <h5 data-tid="project-name">{userTokenData.title}</h5>
+    {#if nonNullish(userTokenData.subtitle)}
+      <span data-tid="project-subtitle" class="description"
+        >{userTokenData.subtitle}</span
+      >
+    {/if}
+  </div>
+</div>
+
+<style lang="scss">
+  h5 {
+    margin: 0;
+  }
+
+  .title-logo-wrapper {
+    display: flex;
+    align-items: center;
+    gap: var(--padding);
+
+    .title-wrapper {
+      display: flex;
+      flex-direction: column;
+      gap: var(--padding-0_5x);
+    }
+  }
+</style>
